requirements.txt
alembic==1.10.3
Flask==2.2.3
Flask_Login==0.6.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
Pillow==9.5.0
Pillow==9.5.0
SQLAlchemy==2.0.9
Werkzeug==2.2.3
WTForms==3.0.1
